Moved Player and FriendsButton classes to own file
