+++ app/src/main/AndroidManifest.xml
@@ -39,7 +39,7 @@
             android:name=".game.VotingPageActivity"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".home.LeaderboardActivity"
+            android:name=".home.leaderboard.LeaderboardActivity"
             android:screenOrientation="portrait" />
         <activity
             android:name=".home.BattleLogActivity"
